Move WorkspaceBuildProgress
