+++ site/src/pages/WorkspacePage/WorkspaceBuildProgress.stories.tsx
