+++ site/src/pages/WorkspacePage/WorkspaceBuildProgress.tsx
